Modified the declaration for doube instead of float

+++ inc/Staple.h
@@ -19,18 +19,18 @@ class Staple
 public:
             Staple();
    virtual ~Staple();
-            Staple(std::string _daysBeforeExpiration, std::string _stapleName, float _staplePrice, unsigned int _stapleQuantity);
+            Staple(std::string _daysBeforeExpiration, std::string _stapleName, double _staplePrice, unsigned int _stapleQuantity);
 
    std::string  numberOfDaysBeforeExpiration() const;
    std::string  name() const;
-   float        price() const;
+   double       price() const;
    unsigned int quantity() const;
 
 /// protected members
 protected:
    std::string  _daysBeforeExpiration;
    std::string  _stapleName;
-   float        _staplePrice;
+   double       _staplePrice;
    unsigned int _stapleQuantity;
 };
